Refactored minor issues in Player.cpp & world.cpp.
To-Do:

Implement platform generation based on the latest platform in the vector.
Validate distances between platforms to ensure correctness.
Ensure that the last platform in the vector isn't disappearing; if it is, continue checking further until a non-disappearing platform is found and validate its distance.
Guarantee that the player will never fail due to a game error.